Add preferred door position logic and remove unnecessary console logs
A new attribute for preferred door position was added in `HouseRoomAssigner` class to handle door placements in a house more efficiently. Additionally, conditions checking if a parent room type or child room type is a hallway were restructured in `WallInfo` class to utilize the added attribute. This would allow the position of a door to be preferred if it is a hallway. Updating of preferred door positions within the hallway ends was handled in `HouseGenerator` class. Furthermore, some unnecessary console logs which were present for debugging purposes have been removed to clean up the code.